requirements.txt
pyspark==3.5.0
pyyaml==6.0.1
pytest==7.4.3
python-dotenv==1.0.0
dataclasses-json==0.6.3
typer==0.9.0
psycopg2-binary==2.9.9  # PostgreSQL driver
pymysql==1.1.0  # MySQL driver
python-dotenv==1.0.0  # For managing environment variables/credentials 